Move implicit resource tests to its own file
