+++ python/ray/tests/BUILD
@@ -179,6 +179,7 @@ py_test_module_list(
     "test_streaming_generator_2.py",
     "test_streaming_generator_3.py",
     "test_scheduling_performance.py",
+    "test_implicit_resource.py",
   ],
   size = "medium",
   tags = ["exclusive", "medium_size_python_tests_k_to_z", "team:core"],
